+++ src/workers/localStorage.js
@@ -2,7 +2,7 @@ import registerPromiseWorker from "promise-worker/register";
 import localforage from "localforage";
 import _ from "lodash";
 import defaultJVMDashboards from "../json/jvm/dashboards.json";
-// import defaultGolangDashboards from "../json/golang/dashboards.json";
+// import defaultGoDashboards from "../json/go/dashboards.json";
 
 registerPromiseWorker(message => {
   // Temporarily disabled